% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3ensemble-package}
\alias{mlr3ensemble}
\alias{mlr3ensemble-package}
\title{mlr3ensemble: mlr3 extension package for ensemble machine learning}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3ensemble}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3ensemble/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

}
